Add per-cobras status handling.
The Status field in the XML file is currently always 1, indicating a
working cobra. In order to remain compatible with that, 1 will remain
"OK" for now.

COBRA_OK_MASK       = 0x0001  # 1 for good, 0 for bad.
COBRA_BROKEN_MASK   = 0x0002  # 1 if the motors do not work
COBRA_INVISBLE_MASK = 0x0004  # 1 if the fiber is not visible